pyproject.toml
[tool.poetry]
name = "tabulate"
version = "1.0.0"
description = "A Slack bot that automatically finds, formats, and displays tabular data as a clean snippet."
authors = ["Ravi Kumar <ravikumar.iitp@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "tabulate", from = "src/main" }
]

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = "1.18.1"
pandas = "2.2.2"
tabulate = "0.9.0"
python-dotenv = "1.0.1"
flask = "3.0.3"
gunicorn = "22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = [
  "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"